requirements.txt
apache-airflow[celery,crypto,devel,password,postgres,ssh,async]
numpy
pandas
requests
